Safari-609.1.11.1/LayoutTests/crypto/subtle/ecdsa-generate-key-expected.txt

Test generating an EC key pair with P-256 using ECDSA algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating a key pair...
PASS keyPair.toString() is '[object Object]'
PASS keyPair.publicKey.type is 'public'
PASS keyPair.publicKey.extractable is true
PASS keyPair.publicKey.algorithm.name is 'ECDSA'
PASS keyPair.publicKey.algorithm.namedCurve is 'P-256'
PASS keyPair.publicKey.usages is ['verify']
PASS keyPair.privateKey.type is 'private'
PASS keyPair.privateKey.extractable is true
PASS keyPair.privateKey.algorithm.name is 'ECDSA'
PASS keyPair.privateKey.algorithm.namedCurve is 'P-256'
PASS keyPair.privateKey.usages is ['sign']
PASS successfullyParsed is true

TEST COMPLETE
